refactor(storage): restore ipc module path
