+++ src/main/storage/ipc.test.ts
@@ -42,7 +42,7 @@ vi.mock('electron', () => ({
 
 const { initDataRoot } = await import('../storage-root')
 const { createStorageCommandOwner } = await import('./command-owner')
-const { registerStorageIpcHandlers } = await import('./electron-ipc')
+const { registerStorageIpcHandlers } = await import('./ipc')
 const { clearMigrationPending, isMigrationPending } = await import('./migration-state')
 const { clearApplicationShutdownTrigger, currentApplicationShutdownTrigger } =
   await import('../application-shutdown-trigger')
